requirements.txt
fastapi==0.98.0
pydantic==1.10.9
uvicorn==0.22.0
PyYAML==6.0
langchain==0.0.208
chromadb==0.3.26
openai==0.27.8
tiktoken==0.4.0
